# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MESH API
  version: 0.3.0
  license: { name: UNLICENSED, identifier: LicenseRef-UNLICENSED }
  description: |
    Contract for the MESH student-collaboration modular monolith. This milestone exposes account,
    profile, self-declared skills, goals, interests, desired skills, weekly availability, GitHub public-repository evidence,
    discovery, connections, project groups, persistent messages, tasks, calendar events, and availability polls.
    GitHub evidence is a limited supporting signal, never proof of expertise. LinkedIn is not collected or scraped.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      operationId: register
      summary: Create a student account and an empty working profile
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Account created
          content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/Conflict' }
  /auth/login:
    post:
      operationId: login
      summary: Exchange email and password for a bearer token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated
          content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /projects/{projectId}:
    get:
      operationId: getProject
      summary: Get one project and its members
      parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
      responses:
        '200': { description: Project, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateProject
      summary: Update a project; owner only
      parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProjectInput' } } }
      responses:
        '200': { description: Updated project, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteProject
      summary: Delete a project; owner only
      parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
      responses:
        '204': { description: Project deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /projects/{projectId}/events:
    get:
      operationId: listProjectEvents
      summary: List project calendar events; project members only
      parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
      responses:
        '200': { description: Calendar events, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/CalendarEvent' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      operationId: createProjectEvent
      summary: Create a project calendar event; project members only
      parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateCalendarEventRequest' } } }
      responses:
        '201': { description: Event created, content: { application/json: { schema: { $ref: '#/components/schemas/CalendarEvent' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /projects/{projectId}/events/{eventId}:
    put:
      operationId: updateProjectEvent
      summary: Update a project calendar event; project members only
      parameters: [ { $ref: '#/components/parameters/ProjectId' }, { $ref: '#/components/parameters/EventId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateCalendarEventRequest' } } }
      responses:
        '200': { description: Event updated, content: { application/json: { schema: { $ref: '#/components/schemas/CalendarEvent' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteProjectEvent
      summary: Delete a project calendar event; project members only
      parameters: [ { $ref: '#/components/parameters/ProjectId' }, { $ref: '#/components/parameters/EventId' } ]
      responses:
        '204': { description: Event deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /projects/{projectId}/availability-polls:
    get:
      operationId: listAvailabilityPolls
      summary: List availability polls and vote totals; project members only
      parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
      responses:
        '200': { description: Availability polls, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/AvailabilityPoll' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      operationId: createAvailabilityPoll
      summary: Create an availability poll with two to eight time options; project members only
      parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateAvailabilityPollRequest' } } }
      responses:
        '201': { description: Poll created, content: { application/json: { schema: { $ref: '#/components/schemas/AvailabilityPoll' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /projects/{projectId}/availability-polls/{pollId}/options/{optionId}/vote:
    put:
      operationId: voteAvailabilityPollOption
      summary: Vote for an open availability option; project members only
      parameters: [ { $ref: '#/components/parameters/ProjectId' }, { $ref: '#/components/parameters/PollId' }, { $ref: '#/components/parameters/OptionId' } ]
      responses:
        '204': { description: Vote recorded; repeated calls are idempotent }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: removeAvailabilityPollVote
      summary: Remove a vote from an open availability option; project members only
      parameters: [ { $ref: '#/components/parameters/ProjectId' }, { $ref: '#/components/parameters/PollId' }, { $ref: '#/components/parameters/OptionId' } ]
      responses:
        '204': { description: Vote removed; succeeds when no vote exists }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /projects/{projectId}/availability-polls/{pollId}/close:
    patch:
      operationId: closeAvailabilityPoll
      summary: Select an option and close an availability poll; project owner only
      parameters: [ { $ref: '#/components/parameters/ProjectId' }, { $ref: '#/components/parameters/PollId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CloseAvailabilityPollRequest' } } }
      responses:
        '200': { description: Closed poll, content: { application/json: { schema: { $ref: '#/components/schemas/AvailabilityPoll' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /projects/{projectId}/availability-polls/{pollId}/schedule:
    post:
      operationId: scheduleAvailabilityPoll
      summary: Create a calendar event from a closed availability poll; project owner only
      parameters: [ { $ref: '#/components/parameters/ProjectId' }, { $ref: '#/components/parameters/PollId' } ]
      responses:
        '200': { description: Scheduled event, content: { application/json: { schema: { $ref: '#/components/schemas/CalendarEvent' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  responses:
    ValidationError:
      description: Request field validation failed
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: Missing, invalid, or expired bearer token
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: A unique value is already in use
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: The signed-in student is not allowed to perform this action
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: The requested resource does not exist or is not accessible
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    PollId:
      name: pollId
      in: path
      required: true
      schema: { type: string, format: uuid }
    OptionId:
      name: optionId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    EventId:
      name: eventId
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    RegisterRequest:
      type: object
      required: [displayName, username, email, password]
      properties:
        displayName: { type: string, minLength: 1, maxLength: 100, example: Asha Patel }
        username: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._-]{2,39}$', example: asha.builds }
        email: { type: string, format: email, maxLength: 254, example: asha@university.edu }
        password: { type: string, format: password, minLength: 8, maxLength: 72 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, example: asha@university.edu }
        password: { type: string, format: password }
    AuthResponse:
      type: object
      required: [token, tokenType, expiresAt, user]
      properties:
        token: { type: string, description: "Send as Authorization: Bearer <token>." }
        tokenType: { type: string, const: Bearer }
        expiresAt: { type: string, format: date-time }
        user: { $ref: '#/components/schemas/Profile' }
...
    Profile:
      type: object
      required: [id, username, displayName, onboardingComplete, skills, goals, interests, desiredSkills, availabilityWindows]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        department: { type: [string, 'null'] }
        yearOfStudy: { type: [integer, 'null'] }
        bio: { type: [string, 'null'] }
        avatarKey: { type: [string, 'null'] }
        availability: { type: [string, 'null'] }
        availabilityTimezone: { type: [string, 'null'], description: IANA timezone for `availabilityWindows`. }
        primaryDomain: { type: [string, 'null'] }
        onboardingComplete: { type: boolean }
        skills: { type: array, items: { $ref: '#/components/schemas/ProfileSkill' } }
        goals: { type: array, items: { $ref: '#/components/schemas/Goal' } }
        interests: { type: array, items: { $ref: '#/components/schemas/Interest' } }
        desiredSkills: { type: array, items: { $ref: '#/components/schemas/Skill' } }
        availabilityWindows: { type: array, items: { $ref: '#/components/schemas/AvailabilityWindow' } }
    Skill:
      type: object
      required: [id, name, category]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        category: { type: string }
    ProfileSkill:
      allOf:
        - $ref: '#/components/schemas/Skill'
        - type: object
          required: [proficiency, evidenceSupported]
          properties:
            proficiency: { type: integer, minimum: 1, maximum: 5, description: Self-declared proficiency. }
            evidenceSupported: { type: boolean, description: "True only when an authorized GitHub public-repository language matches the self-declared skill; it is a supporting signal, not proof of expertise." }
    Goal:
      type: object
      required: [code, name]
      properties:
        code: { type: string, example: HACKATHON }
        name: { type: string, example: Hackathon project }
    Interest:
      type: object
      required: [id, name]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
...
    AvailabilityWindow:
      type: object
      required: [dayOfWeek, startMinute, endMinute]
      properties:
        dayOfWeek: { type: integer, minimum: 1, maximum: 7, description: ISO weekday number; 1 is Monday. }
        startMinute: { type: integer, minimum: 0, maximum: 1439, description: Minutes after midnight in `timezone`. }
        endMinute: { type: integer, minimum: 1, maximum: 1440, description: Must be later than `startMinute`. }
...
    ProjectInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 120, example: Campus accessibility map }
        summary: { type: [string, 'null'], maxLength: 1000, example: Map accessible routes around campus. }
        domain: { type: [string, 'null'], maxLength: 80, example: WEB_DEVELOPMENT }
...
    Project:
      type: object
      required: [id, name, ownerUserId, createdAt, members]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        summary: { type: [string, 'null'] }
        domain: { type: [string, 'null'] }
        ownerUserId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        members: { type: array, items: { $ref: '#/components/schemas/ProjectMember' } }
    ProjectMember:
      type: object
      required: [userId, username, displayName, role]
      properties:
        userId: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        avatarKey: { type: [string, 'null'] }
        role: { type: string, enum: [OWNER, MEMBER] }
...
    CreateCalendarEventRequest:
      type: object
      required: [title, startsAt, endsAt]
      properties:
        title: { type: string, minLength: 1, maxLength: 160 }
        description: { type: [string, 'null'], maxLength: 2000 }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time, description: Must be later than `startsAt`. }
    CalendarEvent:
      type: object
      required: [id, projectId, title, startsAt, endsAt, createdByUserId, createdAt]
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        title: { type: string }
        description: { type: [string, 'null'] }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        createdByUserId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
    CreateAvailabilityPollRequest:
      type: object
      required: [title, options]
      properties:
        title: { type: string, minLength: 1, maxLength: 160 }
        options: { type: array, minItems: 2, maxItems: 8, items: { $ref: '#/components/schemas/AvailabilityPollOptionInput' } }
    AvailabilityPollOptionInput:
      type: object
      required: [startsAt, endsAt]
      properties:
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time, description: Must be later than `startsAt`. }
    CloseAvailabilityPollRequest:
      type: object
      required: [selectedOptionId]
      properties:
        selectedOptionId: { type: string, format: uuid }
    AvailabilityPoll:
      type: object
      required: [id, projectId, title, status, createdByUserId, createdAt, options]
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        title: { type: string }
        status: { type: string, enum: [OPEN, CLOSED, SCHEDULED] }
        selectedOptionId: { type: [string, 'null'], format: uuid }
        createdByUserId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        options: { type: array, items: { $ref: '#/components/schemas/AvailabilityPollOption' } }
    AvailabilityPollOption:
      type: object
      required: [id, startsAt, endsAt, voteCount, votedByMe]
      properties:
        id: { type: string, format: uuid }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        voteCount: { type: integer, minimum: 0 }
        votedByMe: { type: boolean }
    Error:
      type: object
      required: [timestamp, status, error, message, fields, path]
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        message: { type: string }
        fields: { type: object, additionalProperties: { type: string } }
        path: { type: string }